# docs/architecture/openapi.yaml
openapi: 3.1.0
info:
  title: Lynk REST API
  description: Authoritative REST API specification for Lynk, a high-trust campus freelance and gig marketplace built on the Unified Campus Member model.
  version: 1.0.0
  contact:
    name: Lynk Engineering Team
    url: https://github.com/BIJJUDAMA/Lynk
servers:
  - url: http://localhost:8080/api/v1
    description: Local Docker Compose development environment
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /jobs:
    get:
      summary: Search and list campus opportunities
      description: Filter jobs by department, skills, and search keywords.
      security: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: department
          in: query
          required: false
          schema:
            type: string
        - name: skills
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, closed, completed, cancelled]
            default: open
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated job opportunities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'

    post:
      summary: Post a new job opportunity
      description: Creates a new gig. Description capped at 5,000 characters. Skills capped at 25 items. Normalizes deadlines to UTC calendar day.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '201':
          description: Job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Validation error or bounds exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Campus verification pending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /jobs/{id}:
    get:
      summary: Get job details
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update job opportunity
      description: Modifies open job parameters. Closed and cancelled jobs are immutable.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJobRequest'
      responses:
        '200':
          description: Job updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Cannot update closed or cancelled job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to modify job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Cancel job opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Job cancelled successfully
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sAccessToken
      description: SuperTokens HttpOnly session access token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional bearer token for API clients
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Campus verification pending
...
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        created_by:
          type: string
        title:
          type: string
        description:
          type: string
        required_skills:
          type: array
          items:
            type: string
        department:
          type: string
        status:
          type: string
          enum: [open, closed, completed, cancelled]
        deadline:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateJobRequest:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 255
        description:
          type: string
          maxLength: 5000
        required_skills:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 25
        department:
          type: string
        deadline:
          type: string
          format: date-time
    UpdateJobRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
          maxLength: 5000
        required_skills:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 25
        department:
          type: string
        deadline:
          type: string
          format: date-time
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Job'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...